Exception handling

+++ src/main/java/com/phc/dscommerce/services/ProductService.java
@@ -2,14 +2,15 @@ package com.phc.dscommerce.services;
 
 import com.phc.dscommerce.dto.ProductDTO;
 import com.phc.dscommerce.entities.Product;
+import com.phc.dscommerce.exceptions.DatabaseException;
+import com.phc.dscommerce.exceptions.ResourceNotFoundException;
 import com.phc.dscommerce.repositories.ProductRepository;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -21,8 +22,7 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Optional<Product> result = productRepository.findById(id);
-        Product product = result.get();
+        Product product = productRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
         return new ProductDTO(product.getId(), product.getName(), product.getDescription(), product.getPrice(), product.getImgUrl());
     }
 
@@ -39,7 +39,7 @@ public class ProductService {
     }
 
     @Transactional
-    public ProductDTO insertProduct(ProductDTO productDTO){
+    public ProductDTO insertProduct(ProductDTO productDTO) {
         Product product = new Product();
         copyDtoToEntity(productDTO, product);
         productRepository.save(product);
@@ -48,15 +48,25 @@ public class ProductService {
 
     @Transactional
     public ProductDTO updateProduct(Long id, ProductDTO productDTO) {
-        Product product = productRepository.getReferenceById(id);
-        copyDtoToEntity(productDTO, product);
-        productRepository.save(product);
-        return new ProductDTO(product.getId(), product.getName(), product.getDescription(), product.getPrice(), productDTO.getImgUrl());
+        try {
+            Product product = productRepository.getReferenceById(id);
+            copyDtoToEntity(productDTO, product);
+            productRepository.save(product);
+            return new ProductDTO(product.getId(), product.getName(), product.getDescription(), product.getPrice(), productDTO.getImgUrl());
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
-    public void deleteByIdProduct(Long id) {
-        productRepository.deleteById(id);
+    public void deleteProductById(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            productRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO productDTO, Product product) {
